refactor(settings): 拆分 Settings.tsx 巨型文件为 settings/ 子模块 (#331)
Settings.tsx 长到 5206 行，所有顶层常量 + 30 多个子组件挤在单文件里，
定位和改动都很费劲。照 pages/tools/generate/ 的先例拆到 settings/ 子目录：

- settings/constants.ts    类型 / 默认值 / TAB 表 / buildPatch 等纯逻辑
- settings/fields.tsx      SettingsSection / SettingsField / SettingsInput 等 UI 原子
- settings/modelCards.tsx  WD14 / CLTagger / Eval 等模型下载卡片
- settings/sections.tsx    各 tab 的独立 Section 组件
- settings/SystemSection.tsx  系统 tab（版本更新 / 存储 / 服务）整块

主文件 5206 → 1044 行，只留 import + SettingsPage。纯机械搬移，逐字节
不动逻辑 / 注释 / 命名，仅增删 import 与 export。default export 保持不变
（components/SettingsDrawer.tsx 靠 lazy import 依赖）。tsc / vitest / lint /
build 全绿，Settings 仍为独立 lazy chunk。
